requirements.txt
annotated-types==0.7.0
anyio==4.13.0
boto3==1.38.1
botocore==1.38.1
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
filetype==1.2.0
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
s3transfer==0.12.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing_extensions==4.15.0
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
